fix:Adjust delta packet fields, update zone capacity, and fix minor packet buffer issues
- Adjusted DeltaUpdateField layout and optimized field sizes
- Updated zone_capacity configuration for more realistic load handling
- Corrected packet buffer length definitions (normal/big packet size tuning)

+++ ClientCore/PacketHelper/Packet.cs
@@ -215,7 +215,7 @@ namespace ClientCore.PacketHelper
     {
         public ulong zoneInternalID;
         public ushort fieldID;
-        public ulong fieldVal;
+        public uint fieldVal;
     }
 
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
